+++ CHANGES.md
@@ -6,6 +6,7 @@ libcups v3.0.0 (YYYY-MM-DD)
 
 - Added `cupsOAuthGetJWKS` API.
 - Added `httpGetCookieValue` and `httpGetSecurity` APIs.
+- Updated `httpSetCookie` API to support multiple "Set-Cookie:" header values.
 - Updated `ippfind` to use `cupsGetClock` API.
 - Fixed return values of `ippDateToTime` when the timezone isn't GMT.
 - Fixed a potential timing issue with `cupsEnumDests`.